backend/requirements.txt
boto3==1.37.7
Flask==3.1.0
flask_cors==5.0.1
numba==0.61.0
librosa==0.10.2.post1
numpy>=1.24,<2.2
python-dotenv==1.0.1
